doorbell-server/requirements.txt
pygame
paho-mqtt
gpiozero
python-telegram-bot
Flask
picamera